requirements.txt
requests
python-dotenv
resend
anthropic
langgraph
langchain-anthropic
langchain-core
langsmith
pdfplumber
Pillow
trafilatura
beautifulsoup4>=4.12
python-docx
openpyxl
yt-dlp
